feat: changed requiresauthentication process
