+++ backend/package.json
@@ -17,6 +17,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
+    "firebase-admin": "^13.0.2",
     "jsonwebtoken": "^9.0.2",
     "pg": "^8.13.1",
     "ts-node": "^10.9.2",
